requirements.txt
pandas>=2.0.0
numpy>=1.24.0
openai>=1.0.0
anthropic>=0.18.0
sqlalchemy>=2.0.0
prompt-toolkit>=3.0.0
pygments>=2.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotext>=5.0.0
tabulate>=0.9.0
diskcache>=5.6.0
joblib>=1.3.0
click>=8.1.0
python-dotenv>=1.0.0
